Ensure query field of uri exists
